+++ src/Extensions/CodeFormatter.cpp
@@ -18,16 +18,16 @@
 #include "Extensions/CodeFormatter.hpp"
 #include "Core/EventLogger.hpp"
 #include "Core/MessageLogger.hpp"
+#include "Editor/CodeEditor.hpp"
 #include "Settings/SettingsManager.hpp"
 #include "Util/FileUtil.hpp"
-#include "third_party/QCodeEditor/include/QCodeEditor"
 #include <QProcess>
 #include <QTemporaryDir>
 
 namespace Extensions
 {
 
-CodeFormatter::CodeFormatter(QCodeEditor *editor, const QString &lang, bool selectionOnly, bool logOnNoChange,
+CodeFormatter::CodeFormatter(Editor::CodeEditor *editor, const QString &lang, bool selectionOnly, bool logOnNoChange,
                              MessageLogger *log, QObject *parent)
     : QObject(parent), m_editor(editor), m_lang(lang), m_selectionOnly(selectionOnly), m_logOnNoChange(logOnNoChange),
       log(log)